requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
langchain==0.0.335
langchain-openai==0.0.2
openai==1.7.0
sse-starlette==1.6.5
faiss-cpu==1.7.4
python-dotenv==1.0.0
pydantic==2.4.2
tiktoken==0.5.1
